+++ node/__fixtures__/shipping.ts
@@ -1,6 +1,6 @@
-import { Shipping } from '../../clients/shipping'
-import { ADDRESS_TYPES, DELIVERY, PICKUP_IN_STORE } from '../../constants'
+import { AddressType, DELIVERY, PICKUP_IN_STORE } from '../constants'
 import { EMPTY_ORDER_FORM } from './orderForm'
+import { Clients } from '../clients'
 
 const SLA = {
   deliveryIds: [],
@@ -101,7 +101,7 @@ const scheduledDeliverySLA = {
 
 export const deliveryAddress = {
   addressId: 'test',
-  addressType: ADDRESS_TYPES.RESIDENTIAL,
+  addressType: AddressType.RESIDENTIAL,
   city: '',
   complement: '',
   country: 'BRA',
@@ -313,6 +313,8 @@ export const ORDER_FORM_WITH_EMPTY_SHIPPING_DATA = {
   ...EMPTY_ORDER_FORM,
 }
 
-export const shipping = ({
-  shippingAttachmentRequest: jest.fn(),
-} as unknown) as Shipping
+export const clients = ({
+  checkout: {
+    updateOrderFormShipping: jest.fn(),
+  },
+} as unknown) as Clients