+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Refactored internal structure of shipping resolvers.
+
+### Removed
+- Unused `Shipping` client in favor of `Checkout` client.
 
 ## [0.29.2] - 2020-04-01
 ### Fixed